requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
bcrypt==4.0.1
marshmallow==3.20.1
pdfplumber==0.10.3
